refactor: convert graph runtime client to ESM
(cherry picked from commit b54b0d26d1e91ffac5b542d8084e17dce44b5084)
