--- src/VersoBlueprint/Commands/open-target-details.js
@@ -1,18 +1,0 @@
-(function () {
-  function openFromHash() {
-    if (!window.location.hash) return;
-    const id = decodeURIComponent(window.location.hash.slice(1));
-    if (!id) return;
-    const target = document.getElementById(id);
-    if (!target) return;
-    const details = target.matches("details") ? target : target.closest("details");
-    if (details) details.open = true;
-  }
-
-  if (document.readyState === "loading") {
-    document.addEventListener("DOMContentLoaded", openFromHash);
-  } else {
-    openFromHash();
-  }
-  window.addEventListener("hashchange", openFromHash);
-})();
